+++ src/ui/components/article.rs
@@ -7,11 +7,12 @@ use crate::helpers::fmt as fmt_helpers;
 use crate::helpers::strings as str_helpers;
 
 use chrono::{DateTime, Local};
-use tui_realm_stdlib::{Label, Paragraph, Textarea};
+use tui_realm_stdlib::{Label, Paragraph};
+use tui_realm_textarea::TextArea;
 use tuirealm::command::{Cmd, Direction, Position};
 use tuirealm::event::{Key, KeyEvent};
 use tuirealm::props::{
-    Alignment, BorderSides, BorderType, Borders, Color, TextModifiers, TextSpan,
+    Alignment, BorderSides, BorderType, Borders, Color, Style, TextModifiers, TextSpan,
 };
 use tuirealm::{Component, Event, MockComponent, NoUserEvent};
 
@@ -105,38 +106,82 @@ impl Component<Msg, NoUserEvent> for ArticleLink {
     }
 }
 
-#[derive(MockComponent)]
-pub struct ArticleSummary {
-    component: Textarea,
+pub struct ArticleSummary<'a> {
+    component: TextArea<'a>,
 }
 
-impl ArticleSummary {
-    pub fn new(summary: &str) -> Self {
+impl<'a> MockComponent for ArticleSummary<'a> {
+    fn view(&mut self, frame: &mut tuirealm::Frame, area: tuirealm::tui::layout::Rect) {
+        self.component.view(frame, area);
+    }
+
+    fn query(&self, attr: tuirealm::Attribute) -> Option<tuirealm::AttrValue> {
+        self.component.query(attr)
+    }
+
+    fn attr(&mut self, attr: tuirealm::Attribute, value: tuirealm::AttrValue) {
+        self.component.attr(attr, value)
+    }
+
+    fn state(&self) -> tuirealm::State {
+        self.component.state()
+    }
+
+    fn perform(&mut self, cmd: Cmd) -> tuirealm::command::CmdResult {
+        self.component.perform(cmd)
+    }
+}
+
+impl<'a> ArticleSummary<'a> {
+    pub fn new(summary: &str, width: usize) -> Self {
         Self {
-            component: Textarea::default()
+            component: TextArea::new(Self::make_summary_rows(summary, width))
                 .borders(
                     Borders::default()
                         .color(Color::LightCyan)
                         .modifiers(BorderType::Rounded),
                 )
-                .foreground(Color::Reset)
                 .title("Summary", Alignment::Left)
-                .step(4)
-                .highlighted_str("• ")
-                .text_rows(Self::make_summary_rows(summary).as_slice()),
+                .cursor_style(Style::default())
+                .scroll_step(4),
         }
     }
 
     /// Make summary rows
-    fn make_summary_rows(summary: &str) -> Vec<TextSpan> {
+    fn make_summary_rows(summary: &str, width: usize) -> Vec<String> {
         let summary =
             str_helpers::replace_multiple_newlines(summary.trim_matches('\n').trim(), "\n");
         // Split summary by newline
-        summary.split('\n').map(TextSpan::from).collect()
+        let mut lines = Vec::new();
+        let mut summary_it = summary.chars();
+        let mut acc = Vec::new();
+        loop {
+            match summary_it.next() {
+                None => {
+                    // End of summary
+                    // push acc to lines
+                    lines.push(acc.iter().collect());
+                    break;
+                }
+                Some('\n') => {
+                    lines.push(acc.iter().collect());
+                    acc.clear();
+                }
+                Some(ch) => {
+                    acc.push(ch);
+                }
+            }
+            // if width has exceeded, push and clear
+            if acc.len() >= width {
+                lines.push(acc.iter().collect());
+                acc.clear();
+            }
+        }
+        lines
     }
 }
 
-impl Component<Msg, NoUserEvent> for ArticleSummary {
+impl<'a> Component<Msg, NoUserEvent> for ArticleSummary<'a> {
     fn on(&mut self, ev: Event<NoUserEvent>) -> Option<Msg> {
         match ev {
             Event::Keyboard(KeyEvent {