+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Released on ??
 
 - If `content` in RSS feed is populated, use content instead of `summary`
   - This should be preferred since contains much more information than just the summary. Unfortunately sometimes the content is empty, so summary should be used as fallback
+- Changed stdlib `Textarea` to `tui-realm-textarea` component to improve readability of summary
 - Migrated UI to tui-realm 1.7.0
 - Updated dependencies
   - argh `0.1.8`